backend/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# LangGraph and LangChain
langgraph==0.2.16
langchain==0.2.16
langchain-aws==0.1.17
langchain-core==0.2.38
langchain-community==0.2.16

# AWS Services
boto3>=1.34.131,<1.35.0
botocore>=1.34.131,<1.35.0

# Data validation and serialization
pydantic==2.8.2
pydantic-settings==2.4.0

# Document processing
python-docx==1.1.2
PyPDF2==3.0.1

# Vector database
opensearch-py==2.6.0

# Web search (free - no API key needed)
beautifulsoup4==4.14.2
requests==2.32.5

# Additional dependencies for LangGraph/LangChain
langsmith==0.1.121
anthropic==0.34.0

# Testing
pytest==8.3.2
pytest-asyncio==0.24.0
httpx==0.27.0

# Utilities
python-dotenv==1.0.1
typing-extensions==4.12.2